rename: global directory
