+++ src/main/java/com/ll/medium/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.ll.medium.member.service;
 
-import com.ll.medium.DataNotFoundException;
+import com.ll.medium.global.DataNotFoundException;
 import com.ll.medium.member.entity.Member;
 import com.ll.medium.member.entity.MemberCreateForm;
 import com.ll.medium.member.repository.MemberRepository;
